backend/services/document-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
aiofiles==23.2.1
pymongo==4.6.0
pydantic==1.10.18
httpx==0.26.0
starlette==0.27.0
bcrypt==4.0.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.2.1